fix(label): move data-disabled logic to brn label, fix eslint rule
