+++ libs/ui/label/brain/.eslintrc.json
@@ -6,6 +6,7 @@
 			"files": ["*.ts"],
 			"extends": ["plugin:@nx/angular", "plugin:@angular-eslint/template/process-inline-templates"],
 			"rules": {
+				"@angular-eslint/no-host-metadata-property": 0,
 				"@angular-eslint/directive-selector": [
 					"error",
 					{
