+++ libs/ui/label/helm/src/lib/hlm-label.directive.ts
@@ -1,5 +1,4 @@
-import { isPlatformBrowser } from '@angular/common';
-import { computed, Directive, ElementRef, inject, Input, OnInit, PLATFORM_ID, signal } from '@angular/core';
+import { computed, Directive, inject, Input, signal } from '@angular/core';
 import { hlm } from '@spartan-ng/ui-core';
 import { BrnLabelDirective } from '@spartan-ng/ui-label-brain';
 import { cva, VariantProps } from 'class-variance-authority';
@@ -43,28 +42,20 @@ export type LabelVariants = VariantProps<typeof labelVariants>;
 		'[class]': '_computedClass()',
 	},
 })
-export class HlmLabelDirective implements OnInit {
-	private readonly _isBrowser = isPlatformBrowser(inject(PLATFORM_ID));
-	private readonly _element = inject(ElementRef).nativeElement;
-	private _changes?: MutationObserver;
-	private readonly _dataDisabled = signal<boolean | 'auto'>('auto');
+export class HlmLabelDirective {
+	private readonly _brn = inject(BrnLabelDirective, { host: true });
 
-	ngOnInit(): void {
-		if (!this._isBrowser) return;
-		this._changes = new MutationObserver((mutations: MutationRecord[]) => {
-			mutations.forEach((mutation: MutationRecord) => {
-				if (mutation.attributeName !== 'data-disabled') return;
-				// eslint-disable-next-line
-				const state = (mutation.target as any).attributes.getNamedItem(mutation.attributeName)?.value === 'true';
-				this._dataDisabled.set(state ?? 'auto');
-			});
-		});
-		this._changes?.observe(this._element, {
-			attributes: true,
-			childList: true,
-			characterData: true,
-		});
-	}
+	protected readonly _computedClass = computed(() =>
+		hlm(
+			labelVariants({
+				variant: this._variant(),
+				error: this._error(),
+				disabled: this._brn?.dataDisabled() ?? 'auto',
+			}),
+			this._userCls(),
+		),
+	);
+
 	private readonly _variant = signal<LabelVariants['variant']>('default');
 	@Input()
 	set variant(value: LabelVariants['variant']) {
@@ -78,15 +69,9 @@ export class HlmLabelDirective implements OnInit {
 	}
 
 	private readonly _userCls = signal<ClassValue>('');
+
 	@Input()
 	set class(userCls: ClassValue) {
 		this._userCls.set(userCls);
 	}
-
-	protected readonly _computedClass = computed(() =>
-		hlm(
-			labelVariants({ variant: this._variant(), error: this._error(), disabled: this._dataDisabled() }),
-			this._userCls(),
-		),
-	);
 }
